requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
Werkzeug>=2.3.0
Pillow>=9.0.0
markdown>=3.4.0
bleach>=5.0.0